chore: update grammar
The previous version included, for example, ']' as part of the char
token, but also as a literal on the grammar. As both have the same
length it would always default to the one that was declared first in the
lexer, so this wouldn't work.

The current alternative defines extra tokens in the form of rg1 and rg2
for ranges of repetitions and cls for classes of characters (surrounded
by []). This way, the parser would default for those in case of a match
(as they are longer), and would parse the characters as chars otherwise.