fix: update grammar
We can't just use the class as a token, as it's pretty complex to parse
on the lexer, and Python's re module does not support capturing
repeated subgroups (eg r'(.)*').

Instead we define a new token, cls_int for class intervals of
characters, that are parsed from strings like 'a-z'. And we handle
parsing the class on the parser from that.